show data category page

+++ resources/views/admin/category/index.blade.php
@@ -10,12 +10,11 @@
         <div id="page-inner">
             <div class="row">
                 <div class="col-md-12">
-                    <h2>Category List</h2>
-                    <h5>Welcome Jhon Deo , Love to see you back. </h5>
+                    <a href="/admin/category/create" class="btn btn-block btn-info">Add Category</a>
 
                 </div>
                 <div class="col-md-6">
-                    <div class="box">
+                    <div class="col-sm-6">
                         <div class="box-header">
                             <h3 class="box-title">Category List</h3>
                         </div><!-- /.box-header -->
@@ -43,7 +42,7 @@
                                     <td>{{$rs->image}}</td>
                                     <td>{{$rs->status}}</td>
                                     <td><a href="/admin/category/edit/{{$rs->id}}/" class="btn btn-info">Edit</a> </td>
-                                    <td><a href="/admin/category/delete/{{$rs->id}}/"class="btn  btn-danger btn-sm">Delete</a></td>
+                                    <td><a href="/admin/category/delete/{{$rs->id}}/"class="btn  btn-danger btn-sm" onclick="return confirm('Deleting!! Are you sure?')">Delete</a></td>
                                     <td><a href="/admin/category/show/{{$rs->id}}/"class="btn btn-success btn-sm">Show</a></td>
 
                                 </tr>
